+++ CHANGELOG.md
@@ -5,11 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.5.1] - 2024-05-23
+## [unreleased] - 2024-05-23
 
 ### Added
 
-- Add support for multiplying parentheses
+- Add parser support for implicit multiplication (thanks [juca1331](https://github.com/juca1331))
 
 ## [2.5.0] - 2024-04-16